+++ assets/training/finetune_acft_hf_nlp/src/model_selector/model_selector.py
@@ -160,9 +160,6 @@ ACFT_CONFIG = {
                 "model_hf_load_kwargs": {
                     "trust_remote_code": True,
                 },
-                "tokenizer_config": {
-                    "return_token_type_ids": False,
-                },
             },
             "mlflow_save_model_kwargs": {
                 "extra_pip_requirements": ["einops"],
@@ -428,8 +425,7 @@ def model_selector(args: Namespace):
                             for key2 in mlflow_data["flavors"][key]:
                                 if key2 == "generator_config" and args.task_name == "TextGeneration":
                                     generator_config = mlflow_data["flavors"][key]["generator_config"]
-                                    mlflow_ftconf_data.update(
-                                        {
+                                    mlflow_ftconf_data_temp = {
                                             "load_config_kwargs": copy.deepcopy(generator_config),
                                             "mlflow_ft_conf": {
                                                 "mlflow_hftransformers_misc_conf": {
@@ -437,18 +433,17 @@ def model_selector(args: Namespace):
                                                 },
                                             },
                                         }
-                                    )
+                                    mlflow_ftconf_data = deep_update(mlflow_ftconf_data_temp, mlflow_ftconf_data)
                                 elif key2 == "model_hf_load_kwargs":
                                     model_hf_load_kwargs = mlflow_data["flavors"][key]["model_hf_load_kwargs"]
-                                    mlflow_ftconf_data.update(
-                                        {
+                                    mlflow_ftconf_data_temp = {
                                             "mlflow_ft_conf": {
                                                 "mlflow_hftransformers_misc_conf": {
                                                     "model_hf_load_kwargs": copy.deepcopy(model_hf_load_kwargs),
                                                 },
                                             },
                                         }
-                                    )
+                                    mlflow_ftconf_data = deep_update(mlflow_ftconf_data_temp, mlflow_ftconf_data)
                 ft_config_data = deep_update(mlflow_ftconf_data, ft_config_data)
                 logger.info(f"Updated FT config data - {ft_config_data}")
             except Exception: