+++ assets/training/model_management/src/azureml/model/mgmt/processors/factory.py
@@ -13,6 +13,7 @@ from azureml.model.mgmt.processors.transformers.config import (
 )
 from azureml.model.mgmt.processors.pyfunc.config import (
     MMLabDetectionTasks,
+    MMLabTrackingTasks,
     SupportedTasks as PyFuncSupportedTasks,
     SupportedTextToImageModelFamily,
 )
@@ -23,7 +24,9 @@ from azureml.model.mgmt.processors.transformers.convertors import (
     WhisperMLflowConvertor,
 )
 from azureml.model.mgmt.processors.pyfunc.convertors import (
+    BLIP2MLFlowConvertor,
     MMLabDetectionMLflowConvertor,
+    MMLabTrackingMLflowConvertor,
     CLIPMLFlowConvertor,
     StableDiffusionMlflowConvertor,
     StableDiffusionInpaintingMlflowConvertor,
@@ -54,8 +57,13 @@ def get_mlflow_convertor(model_framework, model_dir, output_dir, temp_dir, trans
         elif task == SupportedTasks.AUTOMATIC_SPEECH_RECOGNITION.value:
             return ASRMLflowConvertorFactory.create_mlflow_convertor(model_dir, output_dir, temp_dir, translate_params)
         # Models from Hugging face framework exported in PyFunc mlflow flavor
-        elif task == PyFuncSupportedTasks.ZERO_SHOT_IMAGE_CLASSIFICATION.value:
+        elif task in \
+                [PyFuncSupportedTasks.ZERO_SHOT_IMAGE_CLASSIFICATION.value, PyFuncSupportedTasks.EMBEDDINGS.value]:
             return CLIPMLflowConvertorFactory.create_mlflow_convertor(
+                model_dir, output_dir, temp_dir, translate_params
+            )
+        elif task == PyFuncSupportedTasks.IMAGE_TO_TEXT.value:
+            return BLIP2MLflowConvertorFactory.create_mlflow_convertor(
                 model_dir, output_dir, temp_dir, translate_params
             )
         else:
@@ -64,6 +72,10 @@ def get_mlflow_convertor(model_framework, model_dir, output_dir, temp_dir, trans
         # Models from MMLAB model framework exported in PyFunc mlflow flavor
         if MMLabDetectionTasks.has_value(task):
             return MMLabDetectionMLflowConvertorFactory.create_mlflow_convertor(
+                model_dir, output_dir, temp_dir, translate_params
+            )
+        elif MMLabTrackingTasks.has_value(task):
+            return MMLabTrackingMLflowConvertorFactory.create_mlflow_convertor(
                 model_dir, output_dir, temp_dir, translate_params
             )
         else:
@@ -182,12 +194,38 @@ class CLIPMLflowConvertorFactory(MLflowConvertorFactoryInterface):
         )
 
 
+class BLIP2MLflowConvertorFactory(MLflowConvertorFactoryInterface):
+    """Factory class for BLIP2 model family."""
+
+    def create_mlflow_convertor(model_dir, output_dir, temp_dir, translate_params):
+        """Create MLflow convertor for BLIP2 model."""
+        return BLIP2MLFlowConvertor(
+            model_dir=model_dir,
+            output_dir=output_dir,
+            temp_dir=temp_dir,
+            translate_params=translate_params,
+        )
+
+
 class LLaVAMLflowConvertorFactory(MLflowConvertorFactoryInterface):
     """Factory class for LLaVA model family."""
 
     def create_mlflow_convertor(model_dir, output_dir, temp_dir, translate_params):
         """Create MLflow convertor for LLaVA model."""
         return LLaVAMLFlowConvertor(
+            model_dir=model_dir,
+            output_dir=output_dir,
+            temp_dir=temp_dir,
+            translate_params=translate_params,
+        )
+
+
+class MMLabTrackingMLflowConvertorFactory(MLflowConvertorFactoryInterface):
+    """Factory class for MMTrack video model family."""
+
+    def create_mlflow_convertor(model_dir, output_dir, temp_dir, translate_params):
+        """Create MLflow convertor for vision tasks."""
+        return MMLabTrackingMLflowConvertor(
             model_dir=model_dir,
             output_dir=output_dir,
             temp_dir=temp_dir,