+++ scripts/azureml-assets/CHANGELOG.md
@@ -3,6 +3,14 @@
 
 ### 🐛 Bugs Fixed
 
+## 1.16.12 (2023-10-12)
+### 🐛 Bugs Fixed
+- [#1458](https://github.com/Azure/azureml-assets/pull/1458) Update schema for prompt and benchmark assets
+
+## 1.16.11 (2023-10-09)
+### 🐛 Bugs Fixed
+- [#1428](https://github.com/Azure/azureml-assets/pull/1428) Azcopy fixes to support additional clouds
+
 ## 1.16.10 (2023-10-06)
 ### 🐛 Bugs Fixed
 - [#1411](https://github.com/Azure/azureml-assets/pull/1411) Prevent Config._expand_path from returning directories